pyproject.toml
[project]
tags = ["aleph.im", "cod", "time-series", "network", "data-science"]

[tool.poetry]
name = "fishnet-cod"
version = "1.1.0"
description = "Compute-over-Data (CoD) network for time series data on Aleph.im"
authors = ["Mike Hukiewitz <mike.hukiewitz@robotter.ai>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "src/fishnet_cod"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.30.0"
pandas = "^2.0.1"
PyNaCl = "^1.5.0"
base58 = "^2.1.1"
semver = "^3.0.0"
pydantic = "^1.10.7"
fastapi = "^0.95.1"
aleph-sdk-python = {git = "https://github.com/aleph-im/aleph-sdk-python.git"}
aiohttp = "^3.8.4"
aars = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
